缓存测试增加
