+++ admin-web/src/main/java/org/gourd/hu/admin/controller/TestController.java
@@ -8,7 +8,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.gourd.hu.admin.service.TestService;
 import org.gourd.hu.base.holder.RequestHolder;
 import org.gourd.hu.base.response.BaseResponse;
-import org.gourd.hu.cache.annotation.NoRepeatSubmit;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -22,7 +21,7 @@ import javax.servlet.http.HttpServletRequest;
  *
  * @author gourd.hu
  */
-@Api(tags = "项目测试API", description = "项目测试API" )
+@Api(tags = "项目测试API")
 @RestController
 @RequestMapping("/test")
 @Slf4j
@@ -31,12 +30,6 @@ public class TestController {
     @Autowired
     private TestService testService;
 
-    @GetMapping("/repeat")
-    @ApiOperation(value = "测试重复调用")
-    @NoRepeatSubmit(2)
-    public BaseResponse repeatTest(@RequestParam String param) {
-        return BaseResponse.ok("success,参数："+param);
-    }
 
     @GetMapping("/log")
     @ApiOperation(value = "测试操作日志")
